fix(simulate): keep old simulate on past fonction
