+++ api/services/policy/src/actions/SimulateOnPastAction.ts
@@ -1,19 +1,13 @@
-import { handler, KernelInterfaceResolver } from '@ilos/common';
+import { handler } from '@ilos/common';
 import { Action as AbstractAction } from '@ilos/core';
 import { copyGroupIdAndApplyGroupPermissionMiddlewares, validateDateMiddleware } from '@pdc/provider-middleware';
-import { MemoryMetadataRepository } from './../providers/MemoryMetadataRepositoryProvider';
 
-import { handlerConfig, ParamsInterface, ResultInterface } from '../shared/policy/simulateOnPastGeo.contract';
-import {
-  ParamsInterface as GeoParamsInterface,
-  ResultInterface as GeoResultInterface,
-  signature as geoSignature,
-} from '../shared/territory/findGeoBySiren.contract';
+import { handlerConfig, ParamsInterface, ResultInterface } from '../shared/policy/simulateOnPast.contract';
 
-import { MetadataStore } from '../engine/entities/MetadataStore';
-import { Policy } from '../engine/entities/Policy';
-import { SerializedPolicyInterface, TripRepositoryProviderInterfaceResolver } from '../interfaces';
 import { alias } from '../shared/policy/simulateOn.schema';
+import { TripRepositoryProviderInterfaceResolver, TerritoryRepositoryProviderInterfaceResolver } from '../interfaces';
+import { Policy } from '../engine/entities/Policy';
+import { MetadataStore } from '../engine/entities/MetadataStore';
 
 @handler({
   ...handlerConfig,
@@ -34,57 +28,25 @@ import { alias } from '../shared/policy/simulateOn.schema';
 export class SimulateOnPastAction extends AbstractAction {
   constructor(
     private tripRepository: TripRepositoryProviderInterfaceResolver,
-    private kernel: KernelInterfaceResolver,
+    private territoryRepository: TerritoryRepositoryProviderInterfaceResolver,
   ) {
     super();
   }
 
   public async handle(params: ParamsInterface): Promise<ResultInterface> {
-    // 0 find related com
-    const geoParamsInterface: GeoParamsInterface = {
-      siren: params.territory_insee,
-    };
-    const geoResult: GeoResultInterface = await this.kernel.call<GeoParamsInterface>(geoSignature, geoParamsInterface, {
-      call: {
-        user: {},
-      },
-      channel: {
-        service: handlerConfig.service,
-      },
-    });
-
-    const today = new Date();
-    const dateMinusOneMonth = new Date();
-    dateMinusOneMonth.setMonth(today.getMonth() - 1);
-
-    const policyTemplateOneMonth: SerializedPolicyInterface = {
-      start_date: dateMinusOneMonth,
-      end_date: today,
-      _id: 1000,
-      name: '',
-      status: 'active',
-      handler: '',
-      incentive_sum: 0,
-      territory_id: 0,
-      territory_selector: undefined,
-    };
-
     // 1. Find selector and instanciate policy
-    const policy = await Policy.import(policyTemplateOneMonth);
+    const territory_selector = await this.territoryRepository.findSelectorFromId(params.policy.territory_id);
+    const policy = await Policy.import({ ...params.policy, territory_selector, _id: 1 });
 
     // 2. Start a cursor to find trips
-    const cursor = this.tripRepository.findTripByGeo(
-      geoResult.coms.map((m) => m.insee),
-      policy.start_date,
-      policy.end_date,
-    );
+    const cursor = this.tripRepository.findTripByPolicy(policy, policy.start_date, policy.end_date);
     let done = false;
 
     let carpool_total = 0;
     let carpool_subsidized = 0;
     let amount = 0;
 
-    const store = new MetadataStore(new MemoryMetadataRepository());
+    const store = new MetadataStore();
     do {
       const results = await cursor.next();
       done = results.done;
@@ -103,6 +65,7 @@ export class SimulateOnPastAction extends AbstractAction {
       }
     } while (!done);
 
+    // TODO approximation à éviter
     return {
       trip_subsidized: carpool_subsidized,
       trip_excluded: carpool_total / 2 - carpool_subsidized,
