+++ api/services/policy/src/actions/SimulateOnPastByGeoAction.ts
@@ -1,13 +1,22 @@
-import { handler } from '@ilos/common';
+import { handler, KernelInterfaceResolver } from '@ilos/common';
 import { Action as AbstractAction } from '@ilos/core';
 import { copyGroupIdAndApplyGroupPermissionMiddlewares } from '@pdc/provider-middleware';
 
 import { handlerConfig, ParamsInterface, ResultInterface } from '../shared/policy/simulateOnPastGeo.contract';
 
 import { MetadataStore } from '../engine/entities/MetadataStore';
+import { Policy } from '../engine/entities/Policy';
 import { TripRepositoryProviderInterfaceResolver } from '../interfaces';
 import { alias } from '../shared/policy/simulateOn.schema';
-import { Policy } from '../engine/entities/Policy';
+import { MemoryMetadataRepository } from './../providers/MemoryMetadataRepositoryProvider';
+
+import {
+  ParamsInterface as GeoParamsInterface,
+  ResultInterface as GeoResultInterface,
+  signature as geoSignature,
+} from '../shared/territory/findGeoBySiren.contract';
+
+import { SerializedPolicyInterface } from '../interfaces';
 
 @handler({
   ...handlerConfig,
@@ -20,29 +29,59 @@ import { Policy } from '../engine/entities/Policy';
   ],
 })
 export class SimulateOnPastAction extends AbstractAction {
-  constructor(private tripRepository: TripRepositoryProviderInterfaceResolver) {
+  constructor(
+    private kernel: KernelInterfaceResolver,
+    private tripRepository: TripRepositoryProviderInterfaceResolver,
+  ) {
     super();
   }
 
   public async handle(params: ParamsInterface): Promise<ResultInterface> {
-    /// REWORK NEEDED
-    // 1. Build template policy
-    const policy = await Policy.import({ ...params.policy, territory_selector, _id: 1 });
+    // 0 find related com
+    const geoParamsInterface: GeoParamsInterface = {
+      siren: params.territory_insee,
+    };
+    const geoResult: GeoResultInterface = await this.kernel.call<GeoParamsInterface>(geoSignature, geoParamsInterface, {
+      call: {
+        user: {},
+      },
+      channel: {
+        service: handlerConfig.service,
+      },
+    });
 
-    const end_date: Date = new Date();
-    const start_date: Date = new Date();
-    start_date.setMonth(start_date.getMonth() - 6);
+    const today = new Date();
+    const dateMinusOneMonth = new Date();
+    dateMinusOneMonth.setMonth(today.getMonth() - 1);
 
-    const cursor = this.tripRepository.findTripByGeo(params.territory_insee, start_date, end_date);
-    /// REWORK NEEDED
+    const policyTemplateOneMonth: SerializedPolicyInterface = {
+      start_date: dateMinusOneMonth,
+      end_date: today,
+      _id: 1000,
+      name: '',
+      status: 'active',
+      handler: '',
+      incentive_sum: 0,
+      territory_id: 0,
+      territory_selector: undefined,
+    };
+
+    // 1. Find selector and instanciate policy
+    const policy = await Policy.import(policyTemplateOneMonth);
+
     // 2. Start a cursor to find trips
+    const cursor = this.tripRepository.findTripByGeo(
+      geoResult.coms.map((m) => m.insee),
+      policy.start_date,
+      policy.end_date,
+    );
     let done = false;
 
     let carpool_total = 0;
     let carpool_subsidized = 0;
     let amount = 0;
 
-    const store = new MetadataStore();
+    const store = new MetadataStore(new MemoryMetadataRepository());
     do {
       const results = await cursor.next();
       done = results.done;
@@ -61,7 +100,6 @@ export class SimulateOnPastAction extends AbstractAction {
       }
     } while (!done);
 
-    // TODO approximation à éviter
     return {
       trip_subsidized: carpool_subsidized,
       trip_excluded: carpool_total / 2 - carpool_subsidized,
